Move ApplicationInExcludeList to Rule
