+++ src/OpenTelemetry.AutoInstrumentation.StartupHook/StartupHook.cs
@@ -14,10 +14,8 @@
 // limitations under the License.
 // </copyright>
 
-using System.Diagnostics;
 using System.Reflection;
 using System.Runtime.Versioning;
-using OpenTelemetry.AutoInstrumentation;
 using OpenTelemetry.AutoInstrumentation.Logging;
 using OpenTelemetry.AutoInstrumentation.RulesEngine;
 
@@ -53,27 +51,18 @@ internal class StartupHook
             }
         }
 
-        var applicationName = GetApplicationName();
-        Logger.Information($"StartupHook loaded for application with name {applicationName}.");
-
-        if (IsApplicationInExcludeList(applicationName))
-        {
-            Logger.Information("Application is in the exclusion list. Skipping initialization.");
-            return;
-        }
-
-        Logger.Information("Attempting initialization.");
-
-        LoaderAssemblyLocation = GetLoaderAssemblyLocation();
-
         try
         {
+            LoaderAssemblyLocation = GetLoaderAssemblyLocation();
+
             var ruleEngine = new RuleEngine();
             if (!ruleEngine.Validate())
             {
                 Logger.Error("Rule Engine Failure: One or more rules failed validation. Auto-Instrumentation won't be loaded.");
                 return;
             }
+
+            Logger.Information("Attempting initialization.");
 
             // Creating an instance of OpenTelemetry.AutoInstrumentation.Loader.Startup
             // will initialize Instrumentation through its static constructor.
@@ -116,59 +105,6 @@ internal class StartupHook
         {
             Logger.Error($"Error getting loader directory location: {ex}");
             throw;
-        }
-    }
-
-    private static string GetApplicationName()
-    {
-        try
-        {
-            return AppDomain.CurrentDomain.FriendlyName;
-        }
-        catch (Exception ex)
-        {
-            Logger.Error($"Error getting AppDomain.CurrentDomain.FriendlyName: {ex}");
-            return string.Empty;
-        }
-    }
-
-    private static bool IsApplicationInExcludeList(string applicationName)
-    {
-        return GetExcludedApplicationNames().Contains(applicationName);
-    }
-
-    private static List<string> GetExcludedApplicationNames()
-    {
-        var excludedProcesses = new List<string>();
-
-        var environmentValue = GetEnvironmentVariable("OTEL_DOTNET_AUTO_EXCLUDE_PROCESSES");
-
-        if (environmentValue == null)
-        {
-            return excludedProcesses;
-        }
-
-        foreach (var processName in environmentValue.Split(Constants.ConfigurationValues.Separator))
-        {
-            if (!string.IsNullOrWhiteSpace(processName))
-            {
-                excludedProcesses.Add(processName.Trim());
-            }
-        }
-
-        return excludedProcesses;
-    }
-
-    private static string? GetEnvironmentVariable(string variableName)
-    {
-        try
-        {
-            return Environment.GetEnvironmentVariable(variableName);
-        }
-        catch (Exception ex)
-        {
-            Logger.Error($"Error getting environment variable {variableName}: {ex}");
-            return null;
         }
     }
 }
